fix the shaders path
osgDB::findDataFile was not able to find the desired files and was
returning an empty string. Therefore the shaders were not correctly
loaded and instead of throwing, only a error message was being
displayed.

The unit tests are now loading correctly the shaders and displaying it
on the visualization window